cleanup(ε.1): extract decode_block/types.rs (audio sample trait + tunables)
First slice of the `decode_block.rs` per-stage split planned in
`docs/CLEANUP_2026_05.md` ε. Moves the cross-cutting items (referenced
by ≥2 of the per-stage pieces yet to be carved out) into a new
`mfsk-core/src/ft8/decode_block/types.rs` submodule:

- `AudioSample` trait + impls for `i16` / `i8`
- Tunable constants: `NFFT_SPEC`, `NSTEP`, `NSSY`, `TONE_SPACING_HZ`,
  `SAMPLE_RATE_HZ`, `TX_START_OFFSET_S`, `SYNC_LAG_S_DEFAULT`,
  `NMS_ALPHA`, `DEFAULT_Q_THRESH`, `RATIO_EPS_DEFAULT`
- Runtime tunable fns: `ratio_eps()`, `sync_lag_s()`

`Spectrogram`, `SpecCell`, `FP_SPEC_SHIFT`, `CoarseAcc` stay in the
parent for ε.2 — they belong with `compute_spectrogram` (the
spectrogram stage submodule).

External-caller path preserved: `mfsk_core::ft8::decode_block::*`
re-exports the same public symbols (`AudioSample`, `NFFT_SPEC`,
`DEFAULT_Q_THRESH`) via `pub use types::*` in the parent. The
remaining ex-private items are surfaced internally with `pub(super)`
+ a private `use types::*` in the parent.

Acceptance:
- `cargo check` clean on all 4 feature cells (default, embedded ship
  `alloc ft8 fft-extern fixed-point`, host fixed-point bench, embedded
  `nstep-half`)
- FT8 qso3_busy golden tests pass: WSJT-X AP-off floor, JTDX recall
  floor, AP-on strict-superset, AP-on subtract diag, all 7
  `decode_block::tests::*` lib tests
- `cargo clippy` clean, `cargo fmt --check` clean
- Diff confined to `mfsk-core/src/ft8/decode_block.rs` (-168 / +17)
  + the new `decode_block/types.rs`
